Added a new test object, a 64 bit Go language test. In the new diederen directory. DWARFTEST.sh now runs some tests with that executable. RUNALL.sh has new options to make running a single test-set easier.

+++ RUNALL.sh
@@ -1,5 +1,33 @@
+#!/bin/sh
+#
+# By default runs all 3 test sets.
+# To run just one, add one  option, one of dd ddtodd2 or dd2
 
+dodd=y
+dodd2=y
+doddtodd2=y
 
+if [ $# -eq 0 ]
+then
+  echo "Running all 3 tests"
+else
+  if [ $# -eq 1 ]
+  then
+    dodd=n
+    dodd2=n
+    doddtodd2=n
+    case $1 in
+      dd) dodd=y ;;
+      dd2) dodd2=y ;;
+      ddtodd2) doddtodd2=y ;;
+      *) echo "Incorrect arg, only one of: dd dd2 ddtodd2 allowed." ; exit 1 ;;
+    esac
+  else
+    echo "Only one argument of: dd dd2 ddtodd2 allowed."
+    exit 1
+  fi
+fi
+echo "dd? $dodd dd2? $dodd2 ddtodd2? $doddtodd2"
 
 chkres() {
 if test $1 != 0
@@ -30,27 +58,46 @@ chkfail () {
   rm -f junkck2
 }
 
-rm -f ALLdd 
-rm -f ALLdd2 
-rm -f ALLddtodd2
+if [ $dodd = "y" ]
+then
+  rm -f ALLdd 
+fi
+if [ $dodd2 = "y" ]
+then
+  rm -f ALLdd2 
+fi
+if [ $doddtodd2 = "y" ]
+then
+  rm -f ALLddtodd2
+fi
 start=`date`
 echo "start $start"
-echo begin test dd
-./DWARFTEST.sh dd 2>ALLdd 1>&2
-chkres $? "Failure building test dd results"
-chkfail ALLdd "running test dd"
+if [ $dodd = "y" ]
+then
+  echo begin test dd
+  ./DWARFTEST.sh dd 2>ALLdd 1>&2
+  chkres $? "Failure building test dd results"
+  chkfail ALLdd "running test dd"
+fi
 
-echo begin test dd2
-date
-./DWARFTEST.sh dd2 2>ALLdd2 1>&2
-chkres $? "Failure building test dd2 results"
-chkfail ALLdd2 "running test dd2"
+if [ $doddtodd2 = "y" ]
+then
+  echo begin test ddtodd2
+  date
+  ./DWARFTEST.sh ddtodd2 2>ALLddtodd2 1>&2
+  chkres $? "Failure building test ddtodd2 results"
+  chkfail ALLddtodd2 "running test ddtodd2"
+fi
 
-echo begin test ddtodd2
-date
-./DWARFTEST.sh ddtodd2 2>ALLddtodd2 1>&2
-chkres $? "Failure building test ddtodd2 results"
-chkfail ALLddtodd2 "running test ddtodd2"
+if [ $dodd2 = "y" ]
+then
+  echo begin test dd2
+  date
+  ./DWARFTEST.sh dd2 2>ALLdd2 1>&2
+  chkres $? "Failure building test dd2 results"
+  chkfail ALLdd2 "running test dd2"
+fi
+
 
 endt=`date`
 echo "start $start"
